refactor : TicketBox 클래스 제거(LottoTickets으로 책임 이동)
